refactor: move 5 local provider tests from CLI to SDK (Phase 3.1)
Move agent, command, output-style, plugin, and rule provider tests to
packages/sdk/test/providers/local/. Update imports to use SDK source
paths and update error codes to match SDK conventions:
- E_AGENT_NOT_FOUND, E_COMMAND_NOT_FOUND, etc. -> E_COMPONENT_NOT_FOUND
- E_UNSUPPORTED_TYPE, E_UNSUPPORTED_OP -> E_PROVIDER_UNAVAILABLE

Skip schema validation test for plugin provider (SDK does not validate).
