+++ packages/sdk/test/providers/local/output-style.test.ts
@@ -2,7 +2,7 @@ import { afterEach, beforeEach, describe, expect, test } from 'bun:test'
 import { mkdir, mkdtemp, rm, writeFile } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
-import { LocalOutputStyleProvider } from '../../src/lib/component/provider-output-style'
+import { LocalOutputStyleProvider } from '../../../src/providers/local/output-style'
 
 // ---------------------------------------------------------------------------
 // Temp directory setup
@@ -231,7 +231,7 @@ describe('LocalOutputStyleProvider info', () => {
     expect(result.ok).toBe(false)
     if (result.ok) return
 
-    expect(result.error.code).toBe('E_OUTPUT_STYLE_NOT_FOUND')
+    expect(result.error.code).toBe('E_COMPONENT_NOT_FOUND')
   })
 
   test('returns error for unsupported type', async () => {
@@ -241,6 +241,6 @@ describe('LocalOutputStyleProvider info', () => {
     expect(result.ok).toBe(false)
     if (result.ok) return
 
-    expect(result.error.code).toBe('E_UNSUPPORTED_TYPE')
+    expect(result.error.code).toBe('E_PROVIDER_UNAVAILABLE')
   })
 })
